refactor: extract study autoplay hook
