+++ src/features/study/model/useStudyControls.ts
@@ -1,8 +1,9 @@
 import type { DeckId } from "@/entities/deck";
-import { moveStudySession, planStudySessionAutoPlay, setStudySessionIndex } from "@/entities/study-session";
+import { setStudySessionIndex } from "@/entities/study-session";
 
 import * as React from "react";
 
+import { useAutoPlay } from "./useAutoPlay";
 import type { StudySessionState } from "./useStudySessionState";
 
 interface StudyControlOptions {
@@ -16,22 +17,12 @@ export const useStudyControls = (
   { defaultAutoPlay, cardInterval }: StudyControlOptions
 ) => {
   const [showBackText, setShowBackText] = React.useState(false);
-  const [autoPlay, setAutoPlay] = React.useState(defaultAutoPlay);
   const hideBackText = () => setShowBackText(false);
-  const autoPlayPlan = planStudySessionAutoPlay(sessionState, {
-    enabled: autoPlay,
-    intervalSeconds: cardInterval,
+  const { autoPlay, toggleAutoPlay } = useAutoPlay(sessionState, {
+    defaultAutoPlay,
+    cardInterval,
+    onAdvance: hideBackText,
   });
-  const autoPlaySession = autoPlayPlan?.session;
-  const autoPlayIntervalSeconds = autoPlayPlan?.intervalSeconds;
-
-  React.useEffect(() => {
-    if (autoPlaySession === undefined || autoPlayIntervalSeconds === undefined) return;
-    const timeout = window.setTimeout(() => {
-      if (moveStudySession(autoPlaySession, "next")) setShowBackText(false);
-    }, autoPlayIntervalSeconds * 1000);
-    return () => window.clearTimeout(timeout);
-  }, [autoPlayIntervalSeconds, autoPlaySession]);
 
   const updateIndex = (currentIndex: number): void => {
     if (!setStudySessionIndex(deckId, currentIndex)) return;
@@ -43,7 +34,7 @@ export const useStudyControls = (
     autoPlay,
     hideBackText,
     toggleBackText: () => setShowBackText((visible) => !visible),
-    toggleAutoPlay: () => setAutoPlay((playing) => !playing),
+    toggleAutoPlay,
     updateIndex,
   };
 };
